backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pyjwt==2.8.1
passlib==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6
httpx==0.25.2
aioredis==2.0.1
redis==5.0.1
requests==2.31.0
cors==1.0.1
